fix: filter state management with url query params
